requirements.txt
streamlit==1.46.0
pyrebase4==4.8.0
requests
pandas
